# contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Changqingjing API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/app/scenics/{entryId}/views": {
      "post": {
        "tags": [
          "app-content-controller"
        ],
        "operationId": "recordScenicView",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordScenicViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseScenicViewResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/app/scenics": {
      "get": {
        "tags": [
          "app-content-controller"
        ],
        "operationId": "scenics",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseAppScenicSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/app/scenics/{entryId}": {
      "get": {
        "tags": [
          "app-content-controller"
        ],
        "operationId": "scenic",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAppScenicResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RecordScenicViewRequest": {
        "required": [
          "viewId"
        ],
        "type": "object",
        "properties": {
          "viewId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ApiResponseScenicViewResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ScenicViewResponse"
          }
        }
      },
      "ScenicViewResponse": {
        "type": "object",
        "properties": {
          "viewCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PageQuery": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "maximum": 100,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponsePageResponseAppScenicSummaryResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PageResponseAppScenicSummaryResponse"
          }
        }
      },
      "AppScenicSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string"
          },
          "openStatus": {
            "type": "string",
            "enum": [
              "OPEN",
              "PAUSED"
            ]
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageResponseAppScenicSummaryResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppScenicSummaryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseAppScenicResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AppScenicResponse"
          }
        }
      },
      "AppScenicBlockResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "HEADING",
              "PARAGRAPH",
              "IMAGE"
            ]
          },
          "text": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "altText": {
            "type": "string"
          }
        }
      },
      "AppScenicResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string"
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppScenicBlockResponse"
            }
          },
          "openStatus": {
            "type": "string",
            "enum": [
              "OPEN",
              "PAUSED"
            ]
          },
          "displayName": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "longitude": {
            "type": "number"
          },
          "latitude": {
            "type": "number"
          },
          "coordinateSystem": {
            "type": "string"
          },
          "firstPublishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "viewCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}